[FE-VIEW] HOC auth 수정중 #7 [미완료]
