+++ client/src/components/views/LoginPage/LoginPage.jsx
@@ -1,51 +1,37 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { useDispatch } from "react-redux";
-import { registerUser } from "../../../_actions/user_action";
+import { loginUser } from "../../../_actions/user_action";
 import { useNavigate } from "react-router-dom";
 
-function RegisterPage() {
+const LoginPage = ({ option }) => {
   const dispatch = useDispatch();
   const navigate = useNavigate();
-
   const [Email, setEmail] = useState("");
-  const [Name, setName] = useState("");
   const [Password, setPassword] = useState("");
-  const [ConfirmPassword, setConfirmPassword] = useState("");
 
   const onEmailHandler = (event) => {
     setEmail(event.currentTarget.value);
-  };
-
-  const onNameHandler = (event) => {
-    setName(event.currentTarget.value);
   };
 
   const onPasswordHandler = (event) => {
     setPassword(event.currentTarget.value);
   };
 
-  const onConfirmPasswordHandler = (event) => {
-    setConfirmPassword(event.currentTarget.value);
-  };
-
   const onSubmitHandler = (event) => {
     event.preventDefault();
 
-    if (Password !== ConfirmPassword) {
-      return alert("비밀번호와 비밀번호 확인이 같지 않습니다.");
-    }
-
     let reqBody = {
       email: Email,
-      name: Name,
       password: Password,
     };
 
-    dispatch(registerUser(reqBody)).then((res) => {
-      if (res.payload.registerSuccess) {
-        navigate("/login");
+    dispatch(loginUser(reqBody)).then((res) => {
+      console.log('res =>', res)
+      console.log('loginSuccess =>', res.payload.loginSuccess)
+      if (res.payload.loginSuccess) {
+        navigate("/");
       } else {
-        alert("Error");
+        alert("로그인 실패");
       }
     });
   };
@@ -66,25 +52,15 @@ function RegisterPage() {
       >
         <label>Email</label>
         <input type="email" value={Email} onChange={onEmailHandler} />
-
-        <label>Name</label>
-        <input type="text" value={Name} onChange={onNameHandler} />
-
         <label>Password</label>
         <input type="password" value={Password} onChange={onPasswordHandler} />
-
-        <label>ConfirmPassword</label>
-        <input
-          type="password"
-          value={ConfirmPassword}
-          onChange={onConfirmPasswordHandler}
-        />
 
         <br />
         <button type="submit">Login</button>
       </form>
+      <button type="button" onClick={() => navigate('/register')}>Register</button>
     </div>
   );
-}
+};
 
-export default RegisterPage;
+export default LoginPage;
