+++ middleware/auth.js
@@ -7,7 +7,7 @@ const auth = async (req, res, next) => {
   if (!token) {
     return res.json({
       isAuth: false,
-      message: "로그인이 필요한 서비스 입니다.",
+      message: "토큰이 없습니다.",
     });
   }
 
@@ -23,6 +23,7 @@ const auth = async (req, res, next) => {
     } else {
       req.token = token;
       req.user = user;
+      req.isAuth = true;
 
       next();
     }
